Updated Post Controller and User Controller

+++ FinalProj/Models/Category.cs
@@ -8,7 +8,7 @@ namespace FinalProj.Models
     public class Category
     {
         public int catID { get; set; }
-        public int webID { get; set; }
+        public long webID { get; set; }
         public string title { get; set; }
         public string desc { get; set; }
         public int currentPage { get; set; }
